build: replace powerbi build with blessed tooling
