--- packages/powerbi-heat-streams-2/capabilities.json
@@ -1,229 +1,0 @@
-{
-	"dataRoles": [
-		{
-			"displayName": "Category",
-			"name": "category",
-			"kind": "Grouping",
-			"requiredTypes": [
-				{
-					"text": true
-				}
-			]
-		},
-		{
-			"displayName": "Score",
-			"name": "measure",
-			"kind": "Measure",
-			"requiredTypes": [
-				{
-					"numeric": true
-				}
-			]
-		},
-		{
-			"displayName": "Time",
-			"name": "grouping",
-			"kind": "Grouping"
-		}
-	],
-	"objects": {
-		"data": {
-			"displayName": "Data Settings",
-			"properties": {
-				"isLogScale": {
-					"displayName": "Log-Scale",
-					"description": "Apply a log-scale to the data coloration.",
-					"type": {
-						"bool": true
-					}
-				},
-				"dateAggregation": {
-					"displayName": "Date Buckets",
-					"description": "When a date domain is used, this specifies date-bucket granularity.",
-					"type": {
-						"enumeration": [
-							{ "displayName": "Hourly", "value": "hours" },
-							{ "displayName": "Daily", "value": "days" },
-							{ "displayName": "Monthly", "value": "months" },
-							{ "displayName": "Yearly", "value": "years" }
-						]
-					}
-				},
-				"numericAggregation": {
-					"displayName": "Value Buckets",
-					"description": "When a NUMERIC domain is used, this specifies number-bucket granularity.",
-					"type": {
-						"numeric": true
-					}
-				},
-				"valueMin": {
-					"displayName": "Min Score",
-					"description": "The minimum score, inferred if not present",
-					"type": {
-						"numeric": true
-					}
-				},
-				"valueMax": {
-					"displayName": "Max Score",
-					"description": "The maximum score, inferred if not present",
-					"type": {
-						"numeric": true
-					}
-				},
-				"scoreSplit": {
-					"displayName": "Score Split",
-					"description": "Score Split Point, midpoint between min and max if not present",
-					"type": {
-						"numeric": true
-					}
-				},
-				"sortBy": {
-					"displayName": "Sort By",
-					"description": "Category Sorting Criteria",
-					"type": {
-						"enumeration": [
-							{ "displayName": "Name", "value": "name" },
-							{ "displayName": "Average", "value": "average" },
-							{ "displayName": "Max", "value": "max" },
-							{ "displayName": "Density", "value": "density" },
-							{ "displayName": "Sum", "value": "sum" }
-						]
-					}
-				},
-				"sortInvert": {
-					"displayName": "Sort Invert",
-					"description": "Invert Sorting Criteria",
-					"type": {
-						"bool": true
-					}
-				}
-			}
-		},
-		"rendering": {
-			"displayName": "View Settings",
-			"properties": {
-				"highlightColor": {
-					"displayName": "Highlight Color",
-					"description": "The selected category highlight color value",
-					"type": {
-						"fill": {
-							"solid": {
-								"color": true
-							}
-						}
-					}
-				},
-				"rowHeight": {
-					"displayName": "Row Height",
-					"description": "The height allocated per category row",
-					"type": {
-						"numeric": true
-					}
-				},
-				"categoryTextPercent": {
-					"displayName": "Text %",
-					"description": "Percentage of the visual to allocate to category text",
-					"type": {
-						"numeric": true
-					}
-				},
-				"rowGap": {
-					"displayName": "Row Gap",
-					"description": "Whether to show a small gap between rows",
-					"type": {
-						"bool": true
-					}
-				},
-				"zoomLevel": {
-					"displayName": "Zoom Level",
-					"description": "Data Zoom Level",
-					"type": {
-						"numeric": true
-					}
-				},
-				"numTicks": {
-					"displayName": "Num Axis Ticks",
-					"description": "The number of axis ticks to present",
-					"type": {
-						"numeric": true
-					}
-				},
-				"showValues": {
-					"displayName": "Show Values",
-					"description": "Show Data Values",
-					"type": {
-						"bool": true
-					}
-				},
-				"showCategories": {
-					"displayName": "Show Categories",
-					"description": "Show Category Labels",
-					"type": {
-						"bool": true
-					}
-				},
-				"colorScheme": {
-					"displayName": "Theme",
-					"description": "The data color scheme",
-					"type": {
-						"enumeration": [
-							{ "displayName": "Diverging - Brown - Green", "value": "BrBG" },
-							{ "displayName": "Diverging - Purple - Green", "value": "PRGn" },
-							{ "displayName": "Diverging - Pink - Green", "value": "PiYG" },
-							{ "displayName": "Diverging - Purple - Orange", "value": "PuOr" },
-							{ "displayName": "Diverging - Red - Blue", "value": "RdBu" },
-							{ "displayName": "Diverging - Red - Grey", "value": "RdGy" },
-							{
-								"displayName": "Diverging - Red - Yellow - Blue",
-								"value": "RdYlBu"
-							},
-							{
-								"displayName": "Diverging - Red - Yellow- Green",
-								"value": "RdYlGn"
-							},
-							{ "displayName": "Diverging - Spectral", "value": "Spectral" },
-							{ "displayName": "Sequential - Blues", "value": "Blues" },
-							{ "displayName": "Sequential - Greens", "value": "Greens" },
-							{ "displayName": "Sequential - Greys", "value": "Greys" },
-							{ "displayName": "Sequential - Oranges", "value": "Oranges" },
-							{ "displayName": "Sequential - Purples", "value": "Purples" },
-							{ "displayName": "Sequential - Reds", "value": "Reds" },
-							{ "displayName": "Categorical - 8 Colors", "value": "Set2" },
-							{ "displayName": "Categorical - 9 Colors", "value": "Set1" },
-							{ "displayName": "Categorical - 12 Colors", "value": "Set3" },
-							{ "displayName": "Categorical - 8 Pastels", "value": "Pastel2" },
-							{ "displayName": "Categorical - 9 Pastels", "value": "Pastel1" },
-							{ "displayName": "Categorical - 8 Accent", "value": "Accent" },
-							{ "displayName": "Categorical - 8 Dark", "value": "Dark2" },
-							{ "displayName": "Categorical - 12 Paired", "value": "Paired" }
-						]
-					}
-				}
-			}
-		}
-	},
-	"dataViewMappings": [
-		{
-			"categorical": {
-				"categories": {
-					"for": {
-						"in": "category"
-					}
-				},
-				"dataVolume": 6,
-				"values": {
-					"group": {
-						"by": "grouping",
-						"select": [
-							{
-								"bind": {
-									"to": "measure"
-								}
-							}
-						]
-					}
-				}
-			}
-		}
-	]
-}
