--- packages/powerbi-heat-streams-2/package.json
@@ -1,31 +1,0 @@
-{
-	"name": "visual",
-	"license": "MIT",
-	"scripts": {
-		"pbiviz": "pbiviz",
-		"clean": "rimraf .tmp dist",
-		"start": "pbiviz start",
-		"build": "pbiviz package"
-	},
-	"dependencies": {
-		"@essex/d3-coloring-scales": "^3.0.0",
-		"@essex/react-heat-streams": "^1.1.4",
-		"@types/d3-array": "^2.0.0",
-		"@types/node": "^13.7.0",
-		"@types/react": "^16.9.19",
-		"@types/react-dom": "^16.9.5",
-		"d3-array": "^2.4.0",
-		"debug": "^4.1.1",
-		"lodash": "^4.17.15",
-		"powerbi-models": "^1.3.2",
-		"powerbi-visuals-api": "^2.6.1",
-		"powerbi-visuals-utils-dataviewutils": "^2.2.1",
-		"powerbi-visuals-utils-svgutils": "^2.2.1",
-		"powerbi-visuals-utils-typeutils": "^2.2.1",
-		"react": "^16.12.0",
-		"react-dom": "^16.12.0"
-	},
-	"devDependencies": {
-		"powerbi-visuals-tools": "^3.1.9"
-	}
-}
