--- packages/powerbi-heat-streams-2/pbiviz.json
@@ -1,25 +1,0 @@
-{
-	"visual": {
-		"name": "heat-streams",
-		"displayName": "Heat Streams",
-		"guid": "heatStreams74779FE5319F4655BE62C4C5FD0ABFB1",
-		"visualClassName": "Visual",
-		"description": "Visualize and compare categorical data values over time using heatmap-style color gradients.",
-		"supportUrl": "https://github.com/microsoft/PowerBI-visuals-HeatStreams/issues",
-		"gitHubUrl": "https://github.com/microsoft/PowerBI-visuals-HeatStreams"
-	},
-	"output": "dist/heatstreams.pbiviz",
-	"apiVersion": "2.6.0",
-	"author": {
-		"name": "Microsoft (experimental)",
-		"email": "msrvizsupport@microsoft.com"
-	},
-	"assets": {
-		"icon": "assets/icon.png"
-	},
-	"externalJS": null,
-	"style": "style/visual.less",
-	"capabilities": "capabilities.json",
-	"dependencies": null,
-	"stringResources": []
-}
