--- packages/powerbi-heat-streams-2/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"compilerOptions": {
-		"allowJs": false,
-		"emitDecoratorMetadata": true,
-		"experimentalDecorators": true,
-		"target": "es6",
-		"jsx": "react",
-		"sourceMap": true,
-		"outDir": "./.tmp/build/",
-		"moduleResolution": "node",
-		"declaration": true,
-		"lib": ["es2015", "dom"]
-	},
-	"files": ["./src/visual.ts"]
-}
