+++ packages/powerbi-heat-streams/package.json
@@ -1,18 +1,11 @@
 {
-	"name": "@essex/visual-heat-streams",
-	"private": true,
-	"version": "1.1.4",
+	"name": "visual",
 	"license": "MIT",
-	"main": "src/Visual.ts",
 	"scripts": {
-		"clean": "rimraf dist",
-		"install-certificate": "essex-powerbi-visual-scripts install-certificate",
-		"build": "cross-env NODE_ENV=production essex-powerbi-visual-scripts package",
-		"start": "essex-powerbi-visual-scripts start"
-	},
-	"author": {
-		"name": "Microsoft (experimental)",
-		"email": "msrvizsupport@microsoft.com"
+		"pbiviz": "pbiviz",
+		"clean": "rimraf .tmp dist",
+		"start": "pbiviz start",
+		"build": "pbiviz package"
 	},
 	"dependencies": {
 		"@essex/d3-coloring-scales": "^3.0.0",
@@ -25,6 +18,7 @@
 		"debug": "^4.1.1",
 		"lodash": "^4.17.15",
 		"powerbi-models": "^1.3.2",
+		"powerbi-visuals-api": "^2.6.1",
 		"powerbi-visuals-utils-dataviewutils": "^2.2.1",
 		"powerbi-visuals-utils-svgutils": "^2.2.1",
 		"powerbi-visuals-utils-typeutils": "^2.2.1",
@@ -32,7 +26,6 @@
 		"react-dom": "^16.12.0"
 	},
 	"devDependencies": {
-		"cross-env": "^7.0.0",
 		"powerbi-visuals-tools": "^3.1.9"
 	}
 }
