+++ packages/powerbi-heat-streams/pbiviz.json
@@ -1,24 +1,25 @@
 {
-  "visual": {
-    "name": "heat-streams",
-    "displayName": "Heat Streams",
-    "guid": "heatStreams74779FE5319F4655BE62C4C5FD0ABFB1",
-    "visualClassName": "Visual",
-    "description": "Visualize and compare categorical data values over time using heatmap-style color gradients.",
-    "supportUrl": "",
-    "gitHubUrl": ""
-  },
-  "output": "dist/heatstreams.pbiviz",
-  "apiVersion": "1.11.0",
-  "author": {
-    "name": "Microsoft (experimental)",
-    "email": "msrvizsupport@microsoft.com"
-  },
-  "assets": {
-    "icon": "assets/icon.png"
-  },
-  "style": "style/visual.scss",
-  "capabilities": "capabilities.json",
-  "dependencies": "dependencies.json",
-  "stringResources": []
+	"visual": {
+		"name": "heat-streams",
+		"displayName": "Heat Streams",
+		"guid": "heatStreams74779FE5319F4655BE62C4C5FD0ABFB1",
+		"visualClassName": "Visual",
+		"description": "Visualize and compare categorical data values over time using heatmap-style color gradients.",
+		"supportUrl": "https://github.com/microsoft/PowerBI-visuals-HeatStreams/issues",
+		"gitHubUrl": "https://github.com/microsoft/PowerBI-visuals-HeatStreams"
+	},
+	"output": "dist/heatstreams.pbiviz",
+	"apiVersion": "2.6.0",
+	"author": {
+		"name": "Microsoft (experimental)",
+		"email": "msrvizsupport@microsoft.com"
+	},
+	"assets": {
+		"icon": "assets/icon.png"
+	},
+	"externalJS": null,
+	"style": "style/visual.less",
+	"capabilities": "capabilities.json",
+	"dependencies": null,
+	"stringResources": []
 }
