+++ packages/powerbi-heat-streams/src/Interactions.ts
@@ -3,6 +3,7 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 /* eslint-disable @typescript-eslint/no-var-requires */
+import powerbi from 'powerbi-visuals-api'
 import { ICategory, TimeDomain } from '@essex/react-heat-streams'
 import buildDomainScrub from './data/buildDomainScrub'
 const get = require('lodash/get')
@@ -14,7 +15,7 @@ export default class Interactions {
 		private selectionManager: powerbi.extensibility.ISelectionManager,
 	) {} // tslint:disable-line no-empty
 
-	public async clearSelections() {
+	public async clearSelections(): Promise<void> {
 		logger.info('Handle Clear')
 		await this.selectionManager.clear()
 		this.applyFilter(null)
@@ -25,7 +26,7 @@ export default class Interactions {
 		category: ICategory,
 		multiselect: boolean,
 		dataView: powerbi.DataView,
-	) {
+	): Promise<void> {
 		logger.info('Handle Cat Click', category, multiselect)
 		const selection = this.selectionIdForCategory(category, dataView)
 		await this.selectionManager.select(selection, multiselect)
@@ -35,14 +36,16 @@ export default class Interactions {
 		this.persistSelectedCategories(selectedCategories)
 	}
 
-	public async scrub(bounds: TimeDomain, dv: powerbi.DataView) {
+	public async scrub(bounds: TimeDomain, dv: powerbi.DataView): Promise<void> {
 		logger.info('Handle Scrub', bounds)
 		if (bounds === null || bounds === undefined || +bounds[0] === +bounds[1]) {
 			this.applyFilter(null)
 			return
 		}
-		const column = dv.metadata.columns.find(col => col.roles.grouping)
-		const filter = buildDomainScrub(bounds, column.identityExprs[0])
+		const column = dv.metadata.columns.find(
+			(col: powerbi.DataViewMetadataColumn) => col.roles!.grouping,
+		)
+		const filter = buildDomainScrub(bounds, column!.identityExprs![0])
 		this.applyFilter(filter)
 	}
 
@@ -52,11 +55,14 @@ export default class Interactions {
 	 */
 	public onRestoreSelection(
 		listener: (ids: powerbi.visuals.ISelectionId[]) => any,
-	) {
+	): void {
 		this.selectionManager.registerOnSelectCallback(listener)
 	}
 
-	private selectionIdForCategory(category: ICategory, dv: powerbi.DataView) {
+	private selectionIdForCategory(
+		category: ICategory,
+		dv: powerbi.DataView,
+	): powerbi.visuals.ISelectionId {
 		const categoryColumn = get(dv, 'categorical.categories[0]', [])
 		return this.host
 			.createSelectionIdBuilder()
@@ -64,14 +70,14 @@ export default class Interactions {
 			.createSelectionId()
 	}
 
-	private applyFilter(filter) {
+	private applyFilter(filter: any): void {
 		logger.info(
 			'Date scrubbing not supported yet in tandem w/ category selection. 🤷‍',
 		)
 		// this.host.applyJsonFilter(filter, 'data', 'filter')
 	}
 
-	private persistSelectedCategories(categories) {
+	private persistSelectedCategories(categories: any): void {
 		// This isn't used yet, but PersistProperties fires the update cycle, which lets selections pipe through
 		this.host.persistProperties({
 			merge: [
