+++ packages/powerbi-heat-streams/src/chart/ChartOptions.ts
@@ -2,6 +2,7 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
+import powerbi from 'powerbi-visuals-api'
 import {
 	Colorizer,
 	DivergingScaler,
@@ -19,12 +20,12 @@ import {
 } from './interfaces'
 
 export default class ChartOptions implements IChartOptions {
-	public dataOptions: IVisualDataOptions
-	public renderOptions: IVisualRenderingOptions
-	public data: IChartData
-	public selections: { [key: string]: ICategory }
-	public timeScrub: TimeDomain | null = null
-	public colorizer: Colorizer
+	public _dataOptions: IVisualDataOptions | undefined
+	public _renderOptions: IVisualRenderingOptions | undefined
+	public _data: IChartData | undefined
+	public _selections: Record<string, ICategory> | undefined
+	public _timeScrub: TimeDomain | null | undefined
+	public _colorizer: Colorizer | undefined
 
 	constructor(
 		private converter: DataViewConverter,
@@ -34,11 +35,11 @@ export default class ChartOptions implements IChartOptions {
 	public loadFromDataView(
 		dataView: powerbi.DataView,
 		settings: VisualSettings,
-	) {
-		this.dataOptions = settings.data
-		this.renderOptions = settings.rendering
-		this.timeScrub = this.converter.unpackDomainScrub(dataView)
-		this.data = this.converter.convertDataView(dataView, settings.data)
+	): void {
+		this._dataOptions = settings.data
+		this._renderOptions = settings.rendering
+		this._timeScrub = this.converter.unpackDomainScrub(dataView)
+		this._data = this.converter.convertDataView(dataView, settings.data)
 		const { colorScheme } = this.renderOptions
 		const { isLogScale } = this.dataOptions
 		const { valueMin, valueMax, valueMid } = this
@@ -46,16 +47,39 @@ export default class ChartOptions implements IChartOptions {
 			? new DivergingScaler(valueMin, valueMid, valueMax, isLogScale)
 			: new LinearScaler(valueMin, valueMax, isLogScale)
 
-		this.colorizer = new Colorizer(scaler, colorScheme)
-
+		this._colorizer = new Colorizer(scaler, colorScheme)
 		this.loadSelections(dataView)
+	}
+
+	public get timeScrub(): TimeDomain | null {
+		return this._timeScrub as TimeDomain | null
+	}
+
+	public get selections(): Record<string, ICategory> {
+		return this._selections as Record<string, ICategory>
+	}
+
+	public get data(): IChartData {
+		return this._data as IChartData
+	}
+
+	public get renderOptions(): IVisualRenderingOptions {
+		return this._renderOptions as IVisualRenderingOptions
+	}
+
+	public get dataOptions(): IVisualDataOptions {
+		return this._dataOptions as IVisualDataOptions
+	}
+
+	public get colorizer(): Colorizer {
+		return this._colorizer as Colorizer
 	}
 
 	/**
 	 * Loads the selections from the given dataView
 	 */
-	public loadSelections(dataView: powerbi.DataView) {
-		this.selections = this.converter.unpackSelectedCategories(dataView)
+	public loadSelections(dataView: powerbi.DataView): void {
+		this._selections = this.converter.unpackSelectedCategories(dataView)
 	}
 
 	public get width(): number {
@@ -80,7 +104,7 @@ export default class ChartOptions implements IChartOptions {
 			: this.data.valueDomain[1]
 	}
 
-	public get valueMid() {
+	public get valueMid(): number {
 		const scoreSplit = this.dataOptions.scoreSplit
 		return scoreSplit !== null && scoreSplit !== undefined
 			? scoreSplit
