+++ packages/powerbi-heat-streams/style/visual.less
