+++ packages/powerbi-heat-streams/tsconfig.json
@@ -1,3 +1,15 @@
 {
-  "extends": "../../tsconfig.json"
+	"compilerOptions": {
+		"allowJs": false,
+		"emitDecoratorMetadata": true,
+		"experimentalDecorators": true,
+		"target": "es6",
+		"jsx": "react",
+		"sourceMap": true,
+		"outDir": "./.tmp/build/",
+		"moduleResolution": "node",
+		"declaration": true,
+		"lib": ["es2015", "dom"]
+	},
+	"files": ["./src/visual.ts"]
 }
